wedding_api/requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
gunicorn==19.8.1
pytz==2018.4
